integrate login with application
