+++ app/src/main/java/com/example/recloth/MainActivity.kt
@@ -5,18 +5,14 @@ import android.util.Log
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.Button
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
 import androidx.credentials.CredentialManager
 import androidx.credentials.CustomCredential
 import androidx.credentials.GetCredentialRequest
 import androidx.lifecycle.lifecycleScope
+import androidx.navigation.NavController
+import androidx.navigation.compose.rememberNavController
+import com.example.recloth.navigation.AppNavigation
+import com.example.recloth.navigation.Screen
 import com.example.recloth.ui.theme.ReClothTheme
 import com.google.android.libraries.identity.googleid.GetGoogleIdOption
 import com.google.android.libraries.identity.googleid.GoogleIdTokenCredential
@@ -35,35 +31,35 @@ class MainActivity : ComponentActivity() {
         enableEdgeToEdge()
         setContent {
             ReClothTheme {
-                LoginScreen(onSignInClick = {
-                    signIn()
-                })
+                val navController = rememberNavController()
+                AppNavigation(navController = navController,
+                    signIn = {
+                        signIn(navController)
+                    })
+//                LoginScreen(onSignInClick = {
+//                    signIn()
+//                })
             }
         }
     }
 
-    private fun firebaseAuthWithGoogle(idToken: String) {
-        // Before using this, make sure you've added the Firebase Auth dependency to build.gradle
-        // implementation 'com.google.firebase:firebase-auth-ktx'
+    private fun firebaseAuthWithGoogle(idToken: String,navController: NavController) {
         val credential = GoogleAuthProvider.getCredential(idToken, null)
         FirebaseAuth.getInstance().signInWithCredential(credential)
             .addOnCompleteListener(this) { task ->
                 if (task.isSuccessful) {
-                    // Sign in success!
                     Log.d("MainActivity", "Firebase sign-in success.")
                     val firebaseUser = FirebaseAuth.getInstance().currentUser
                     Log.d("MainActivity", "User UID: ${firebaseUser?.uid}")
-
                     // TODO: Navigate to your main app screen (AppNavigation) now.
-
+                    navController.navigate(Screen.Landing.route)
                 } else {
-                    // If sign in fails, display a message to the user.
                     Log.w("MainActivity", "Firebase sign-in failed.", task.exception)
                 }
             }
     }
 
-    private fun signIn() {
+    private fun signIn(navController: NavController) {
         val googleIdOption = GetGoogleIdOption.Builder()
             .setFilterByAuthorizedAccounts(false)
             .setServerClientId(getString(R.string.default_web_client_id))
@@ -77,39 +73,26 @@ class MainActivity : ComponentActivity() {
                     request = request,
                     context = this@MainActivity
                 )
-                handleSignInSuccess(result.credential)
+                handleSignInSuccess(result.credential, navController)
             } catch (e: Exception) {
                 Log.e("MainActivity", "Sign-in failed", e)
             }
         }
     }
 
-    private fun handleSignInSuccess(credential: androidx.credentials.Credential) {
+    private fun handleSignInSuccess(credential: androidx.credentials.Credential,navController: NavController) {
         if (credential is CustomCredential && credential.type == GoogleIdTokenCredential.TYPE_GOOGLE_ID_TOKEN_CREDENTIAL) {
             try {
                 val googleIdTokenCredential = GoogleIdTokenCredential.createFrom(credential.data)
                 val googleIdToken = googleIdTokenCredential.idToken
                 Log.d("MainActivity", "Successfully got Google ID token: $googleIdToken")
                 // TODO: Pass this token to your Firebase authentication function
-                firebaseAuthWithGoogle(googleIdToken)
+                firebaseAuthWithGoogle(googleIdToken,navController)
             } catch (e: Exception) {
                 Log.e("MainActivity", "Failed to create credential from data", e)
             }
         } else {
             Log.w("MainActivity", "Credential is not of type Google ID!")
-        }
-    }
-}
-
-@Composable
-fun LoginScreen(onSignInClick: () -> Unit) {
-    Column(
-        modifier = Modifier.fillMaxSize(),
-        verticalArrangement = Arrangement.Center,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        Button(onClick = onSignInClick) {
-            Text("Sign in with Google")
         }
     }
 }
